Producto dto

+++ fotocopiadora_backend/src/main/java/com/example/fotocopiadora_backend/Dto/Producto/PrecioFotocopiaDto.java
@@ -16,5 +16,5 @@ public class PrecioFotocopiaDto {
     private Integer minimo;
 
     @NotNull(message = "El precio unitario no puede estar vacío")
-    private float precioUnitario;
+    private double precioUnitario;
 }
